backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.9.1
python-dotenv==1.0.0
google-generativeai==0.3.2
redis==5.0.1
python-multipart==0.0.6
httpx==0.27.2
websockets>=13.0
